% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(
  parameters,
  init_vals,
  model,
  sadx,
  sady,
  verbose,
  theta,
  alpha_x,
  alpha_y
)
}
\arguments{
\item{parameters}{\code{parameters} corresponds to a vector of parameter
values depending on the provided model: \cr
- model: 'D0' \code{parameters} = c(theta, alpha) \cr
- model: "D1" \code{parameters} = c(theta, alpha X, alpha Y) \cr}

\item{init_vals}{\code{init_vals} corresponds to a vector of parameter values
in which to start the Maxmimum Likelihood algorithm, depending on the provided
model: \cr
- model: "D0" \code{parameters} = c(theta, alpha) \cr
- model: "D1" \code{parameters} = c(theta, alpha X, alpha Y) \cr}

\item{model}{The chosen model to calculate the likelihood for,
please note that the vector of parameters should contain the corresponding
parameters in the right order. The user can pick one of these models:\cr
-  "D0" \cr
-  "D1" \cr}

\item{sadx}{The Species Abundance Distribution of guild X}

\item{sady}{The Species Abundance Distribution of guild Y}

\item{verbose}{TRUE/FALSE flag, indicates whether intermediate output is
shown on screen}

\item{theta}{Fundamental biodiversity number theta}

\item{alpha_x}{Dispersal Ability of Guild X}

\item{alpha_y}{Dispersal Ability of Guild Y}
}
\description{
Default parameter documentation
}
\keyword{internal}
