% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generate.ESF}
\alias{generate.ESF}
\title{Generate community data under the standard neutral model of
biodiversity, using the urn scheme as described in Etienne 2005}
\usage{
generate.ESF(theta, I, J)
}
\arguments{
\item{theta}{Fundamental biodiversity number theta}

\item{I}{Fundamental dispersal number I}

\item{J}{total number of individuals in the local community}
}
\value{
Vector containing the unlabeled species abundances in the local
community
}
\description{
This function generates community data under the standard
neutral model of biodiversity, using the urn scheme as described in
Etienne 2005
}
\details{
generate under the Etienne Sampling Function
}
\examples{
generate.ESF(theta = 42, I = 10, J = 2000)
}
\references{
Etienne, R.S. (2005). A new sampling formula for neutral
biodiversity. Ecology Letters, 8(3), 253-260.
}
\author{
Thijs Janzen & Bart Haegeman
}
