% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generate.Guilds.Cond}
\alias{generate.Guilds.Cond}
\title{Generate Artificial data under the GUILDS model, conditioned on
Guild size}
\usage{
generate.Guilds.Cond(theta, alpha_x, alpha_y, JX, JY)
}
\arguments{
\item{theta}{Fundamental biodiversity number theta}

\item{alpha_x}{Dispersal Ability of Guild X}

\item{alpha_y}{Dispersal Ability of Guild Y}
}
\value{
\item{guildX}{Vector containing the unlabeled abundances of species
in Guild X}
\item{guildY}{Vector containing the unlabeled abundances of species in
Guild Y}
}
\description{
Using this function it is possible to generate a community
dataset consisting of two separate abundance vectors for each guild, where
the data generated adhere to the Guilds model. Data generated is conditioned
on guild size.
}
\details{
conditional on size, generate artificial data under the guilds model
}
\examples{
generate.Guilds.Cond(theta = 200,
                     alpha_x = 0.005,
                     alpha_y = 0.001,
                     JX = 15000,
                     JY = 5000);
}
\author{
Thijs Janzen
}
