% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESF_function.R
\name{logLikelihood.ESF}
\alias{logLikelihood.ESF}
\title{Likelihood of the Etienne sampling formula}
\usage{
logLikelihood.ESF(theta, m, abund)
}
\arguments{
\item{theta}{Parameter value for the fundamental biodiversity number theta}

\item{m}{Parameter value for migration}

\item{abund}{Vector containing abundance data}
}
\value{
loglikelihood
}
\description{
This function calculates the likelihood of the Etienne Sampling
Formula, provided abundance data and parameter values.
}
\details{
loglikelihood following the Etienne Sampling formula
}
\examples{
A <- c(1, 1, 1, 3, 5, 8)   # Artificial abundance dataset
LL <- logLikelihood.ESF(theta = 7, m = 0.1, abund = A)
}
\references{
Etienne, R.S. (2005). A new sampling formula for neutral
biodiversity. Ecology Letters, 8(3), 253-260.
}
\author{
Thijs Janzen
}
