% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{preston_plot}
\alias{preston_plot}
\title{Barplot in Preston style of an abundance dataset}
\usage{
preston_plot(abund, expected, ...)
}
\arguments{
\item{abund}{vector containing the number of individuals per species}

\item{expected}{vector containing the expected number of species per octave}

\item{...}{further graphical arguments that can be passed to \code{barplot()}}
}
\description{
This function first sorts abundances into octaves, and then
plots the resulting distribution.
}
\details{
make a preston plot
}
\examples{
theta = 10
m = 0.1
J = 1000
I = m * (J - 1) / (1 - m)
abund <- generate.ESF(theta, I, J)
par(mfrow = c(1,2))
preston_plot(abund)
abund.expect <- expected.SAD(theta, m, J)
preston_plot(abund, abund.expect)
}
\author{
Thijs Janzen
}
