\name{GUILDS-internal}
\title{Internal Guilds functions}
\alias{simplex}
\alias{rho}
\alias{polyaeggenberger}
\alias{logLikguilds}
\alias{localComm}
\alias{getpx}
\alias{evaluateLogLik}
\alias{esf_local}
\alias{conditional.LogLik}
\alias{calc_sum_kda}
\alias{calc_conditional}
\alias{evaluate_cond_lik}
\alias{calcKDA}
\alias{pm_sad}
\alias{pm_sadaux}
\alias{draw_local_cond}
\alias{draw_local}
\alias{sort_aux}
\alias{generate.ZSM}
\alias{octave_index}
\alias{preston_sort}
\description{Internal Guilds functions}
\details{These are not to be called by the user}
\keyword{internal}