\name{GUILDS-internal}
\title{Internal Guilds functions}
\alias{simplex}
\alias{rho}
\alias{PolyaEggenberger}
\alias{logLikguilds}
\alias{localComm}
\alias{getPX}
\alias{evaluateLogLik}
\alias{ESF_local}
\alias{conditional.LogLik}
\alias{calcSumKDA2}
\alias{calcConditional}
\alias{evaluateCondLik}
\alias{calcKDA}
\alias{pm_sad}
\alias{pm_sadaux}
\alias{drawLocalCond}
\alias{drawLocal}
\alias{prestonSort}
\alias{generate.ZSM}


\description{Internal Guilds functions}
\details{These are not to be called by the user}
\keyword{internal}